requirements.txt
selenium==4.15.2
python-dotenv==1.0.1
requests
bs4